+++ R/purge.R
@@ -4,8 +4,9 @@
 #' for rendering the R Markdown document, by running
 #' \code{docker stop} and \code{docker rm} commands.
 #'
-#' @param input The YAML file (default output of \code{\link{render_docker}})
-#' storing the information of the container to be stopped and removed.
+#' @param input_yml The YAML file (default output of
+#' \code{\link{render_docker}}) storing the information
+#' of the container to be stopped and removed.
 #'
 #' @importFrom yaml yaml.load_file
 #'
@@ -17,14 +18,14 @@
 #' \dontrun{
 #' purge_container("liftr-minimal.docker.yml")}
 
-purge_container = function(input) {
+purge_container = function(input_yml) {
 
-  if (is.null(input))
+  if (is.null(input_yml))
     stop('missing input file')
-  if (!file.exists(normalizePath(input)))
+  if (!file.exists(normalizePath(input_yml)))
     stop('input file does not exist')
 
-  lst = yaml.load_file(normalizePath(input))
+  lst = yaml.load_file(normalizePath(input_yml))
   container_name = lst$'container_name'
 
   # TODO: needs exception handling
@@ -39,8 +40,9 @@ purge_container = function(input) {
 #' for rendering the R Markdown document by
 #' running \code{docker rmi} commands.
 #'
-#' @param input The YAML file (default output of \code{\link{render_docker}})
-#' storing the information of the image to be removed.
+#' @param input_yml The YAML file (default output of
+#' \code{\link{render_docker}}) storing the information
+#' of the image to be removed.
 #'
 #' @importFrom yaml yaml.load_file
 #'
@@ -52,14 +54,14 @@ purge_container = function(input) {
 #' \dontrun{
 #' purge_image("liftr-minimal.docker.yml")}
 
-purge_image = function(input) {
+purge_image = function(input_yml) {
 
-  if (is.null(input))
+  if (is.null(input_yml))
     stop('missing input file')
-  if (!file.exists(normalizePath(input)))
+  if (!file.exists(normalizePath(input_yml)))
     stop('input file does not exist')
 
-  lst = yaml.load_file(normalizePath(input))
+  lst = yaml.load_file(normalizePath(input_yml))
   image_name = lst$'image_name'
 
   # TODO: needs exception handling